Add data types reflecting the contents of each table
Each table has an associated data type where each table attribute is
mapped to an object property. The type of each property should be an
appropriately strict JS type. For example, there's no need for dates in
JS to be represented by strings when JS has a strict `Date` type.